fix: return type of MySQLResultSetView::get_affected_rows

+++ include/coke/mysql_utils.h
@@ -157,7 +157,7 @@ public:
         return std::string_view(info, result_set->info_len);
     }
 
-    int get_affected_rows() const { return result_set->affected_rows; }
+    unsigned long long get_affected_rows() const { return result_set->affected_rows; }
     int get_warnings() const { return result_set->warning_count; }
     unsigned long long get_insert_id() const { return result_set->insert_id; }
